Fixing a few bugs in Stackdriver metrics exporter
- Small refactoring
- Assinging default values for time intervals
