+++ src/Samples/TestStackdriver.cs
@@ -33,7 +33,11 @@
 
         internal static object Run(string projectId)
         {
-            var exporter = new StackdriverExporter(projectId, Tracing.ExportComponent, Stats.ViewManager);
+            var exporter = new StackdriverExporter(
+                projectId, 
+                null,
+                //Tracing.ExportComponent,
+                Stats.ViewManager);
             exporter.Start();
 
             ITagContextBuilder tagContextBuilder = tagger.CurrentBuilder.Put(FrontendKey, TagValue.Create("mobile-ios9.3.5"));
